Update delete path

+++ src/Controller/PinsController.php
@@ -47,7 +47,7 @@ class PinsController extends AbstractController
 
 
     /**
-     * @Route("/pins/{id<[0-9]+>}", name="app_pins_show"), methods="GET"
+     * @Route("/pins/{id<[0-9]+>}", name="app_pins_show", methods="GET")
      */
     public function show(Pin $pin):Response
     {
@@ -55,7 +55,7 @@ class PinsController extends AbstractController
     }
 
     /**
-     * @Route("/pins/{id<[0-9]+>}/edit", name="app_pins_edit"), methods="GET|PUT"
+     * @Route("/pins/{id<[0-9]+>}/edit", name="app_pins_edit", methods="GET|PUT")
      */
     public function edit(Request $request,Pin $pin,EntityManagerInterface $em ):Response
     {
@@ -78,13 +78,14 @@ class PinsController extends AbstractController
     }
 
     /**
-     * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete"), methods="DELETE"
+     * @Route("/pins/{id<[0-9]+>}", name="app_pins_delete", methods="DELETE")
      */
-    public function delete(Pin $pin,EntityManagerInterface $em ):Response
+    public function delete(Request $request, Pin $pin,EntityManagerInterface $em ):Response
     {
-        $em->remove($pin);
-        $em->flush();
-
+        if($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $request->request->get('csrf_token'))){
+            $em->remove($pin);
+            $em->flush();
+        }
         return $this->redirectToRoute('app_home');
     }
 }
